+++ equalsverifier-test-jpms/src/test/java/module-info.java
@@ -4,9 +4,4 @@ open module equalsverifier_jpms_test {
 
     requires org.junit.jupiter.api;
     requires org.assertj.core;
-
-    requires java.desktop;
-    requires java.naming;
-    requires java.rmi;
-    requires java.sql;
 }