+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `forExamples(T red, T blue)` as an alternative to `forClass()` or `forPackage()`. Give it two instances of a type, and it will test that type using the values already present in the fields of the instances. This way, you don't have to use `withPrefabValues` anymore.
+- Prefab value for `java.util.HexFormat`. ([Issue 981](https://github.com/jqno/equalsverifier/issues/981))
 
 ### Changed