+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - EqualsVerifier now requires Java 17.
 - EqualsVerifier is now properly modularized according to the JPMS.
+- Improved error message when ClassCastException is encountered.
 
 ### Removed