+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Exception when superclass of class under test has field with the same name but different type. ([Issue 1056](https://github.com/jqno/equalsverifier/issues/1056))
+- Bug in `withPrefabValueForField` where fields in the superclass of the class under test is ignored.
 
 ## [3.19.1] - 2025-02-17