+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.19.2] - 2025-03-14
+
 ### Fixed
 
 - Exception when superclass of class under test has field with the same name but different type. ([Issue 1056](https://github.com/jqno/equalsverifier/issues/1056))
@@ -1340,7 +1342,8 @@ Please don't use version 1.3; [it's a broken release](https://jqno.nl/post/2013/
 
 You can now use EqualsVerifier!
 
-[unreleased]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.19.1...HEAD
+[unreleased]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.19.2...HEAD
+[3.19.2]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.19.1...equalsverifier-3.19.2
 [3.19.1]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.19...equalsverifier-3.19.1
 [3.19]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.18.2...equalsverifier-3.19
 [3.18.2]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.18.1...equalsverifier-3.18.2